# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DataLogicEngine API
  description: >
    Supported REST integration contract for the local-first DataLogicEngine
    desktop application. Internal desktop-only routes may exist outside this
    external integration surface. The live app uses single-owner OS-level desktop auth, Flask
    sessions, CSRF protection, and copy-once client keys for the explicitly
    versioned gateway surface. Removed web login/register/MFA/SSO routes are intentionally
    not part of this contract.
  version: 4.4.0
  contact:
    name: DataLogicEngine Support
    url: https://github.com/kherrera6219/DataLogicEngine/issues
servers:
  - url: http://localhost:5000/api/v1
    description: Local development or Electron loopback API
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Flask-Login session cookie established by desktop auto-login.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token issued by GET /auth/csrf-token for mutating requests.
    desktopTrustHeader:
      type: apiKey
      in: header
      name: X-DataLogic-Desktop
      description: Desktop loopback trust header used by Electron challenge/auto-login.
    clientBearer:
      type: http
      scheme: bearer
      bearerFormat: ukg_<copy-once-client-key>
    clientApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Copy-once client key. Bearer authentication is preferred.
...
paths:
...
  /memory/review:
    get:
      summary: Review owner memory by trust state
      tags: [Memory]
      security:
        - sessionCookie: []
      responses:
        "200": {description: Working and validated memory review.}
  /memory/export:
    get:
      summary: Export integrity-protected owner memory
      tags: [Memory]
      security:
        - sessionCookie: []
      responses:
        "200": {description: Versioned memory export.}
  /memory/compact:
    post:
      summary: Compact owner memory
      tags: [Memory]
      security:
        - sessionCookie: []
          csrfToken: []
      responses:
        "200": {description: Compaction result.}
  /memory/recover:
    post:
      summary: Recover owner memory from its integrity-checked backup
      tags: [Memory]
      security:
        - sessionCookie: []
          csrfToken: []
      responses:
        "200": {description: Recovery result.}
  /memory/{vertex_id}:
    delete:
      summary: Delete one owner memory vertex
      tags: [Memory]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - name: vertex_id
          in: path
          required: true
          schema: {type: string}
      responses:
        "200": {description: Deletion result.}
...